Add d3-examples bar-chart-sides-style
If bar's value is equal or greater than 0 , then the bar will be
displayed on the right side, its label will be displayed on the Y axis'
left side.
If bar's value is less than 0, than the bar will be displayed on the
left side, its label will be displayed on the Y axis' right side.